requirements.txt
streamlit
google-genai
Pillow
streamlit-stl